backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-mptt>=0.15
psycopg2-binary>=2.9
django-environ>=0.11
django-redis>=5.4
django-cors-headers>=4.3
python-json-logger>=2.0
gunicorn>=21.2
